refactor: sub-clients + README boost
