+++ lib/client/client.ts
@@ -1,18 +1,25 @@
 import { PublicKey } from "@solana/web3.js";
 import { DaoAccount, DaoWithTokens, TokenWithBalance } from "../types";
-import { Proposal, ProposalWithVaults } from "../types/proposals";
+import { ProposalWithVaults } from "../types/proposals";
 
 export interface FutarchyClient {
+  daos: FutarchyDaoClient;
+  proposals: FutarchyProposalsClient;
+  balances: FutarchyBalancesClient;
+}
+
+export interface FutarchyDaoClient {
   fetchAllDaos(): Promise<DaoWithTokens[]>;
   fetchDao(daoAddress: string): Promise<DaoWithTokens | undefined>;
+}
+
+export interface FutarchyProposalsClient {
+  fetchProposals(dao: DaoAccount): Promise<ProposalWithVaults[]>;
+}
+
+export interface FutarchyBalancesClient {
   fetchMainTokenWalletBalances(
     dao: DaoWithTokens,
     ownerWallet: PublicKey
-  ): Promise<TokenWithBalance[]>;
-  fetchProposals(dao: DaoAccount): Promise<ProposalWithVaults[]>;
-  fetchAllConditionalTokenWalletBalances(
-    dao: DaoWithTokens,
-    ownerWallet: PublicKey,
-    proposals: ProposalWithVaults[]
   ): Promise<TokenWithBalance[]>;
 }
