+++ lib/client/indexer/dao.ts
@@ -1,0 +1,30 @@
+import { Program, Provider } from "@coral-xyz/anchor";
+import {
+  AutocratProgram,
+  DaoAccount,
+  DaoWithTokens,
+  ProgramVersion,
+} from "../../types";
+import { FutarchyDaoClient } from "../client";
+import { enrichTokenMetadata } from "../../tokens";
+
+export class FutarchyIndexerDaoClient implements FutarchyDaoClient {
+  constructor() {}
+  async fetchAllDaos(): Promise<DaoWithTokens[]> {
+    return [];
+  }
+  async fetchDao(daoAddress: string): Promise<DaoWithTokens | undefined> {
+    return undefined;
+  }
+  private async fetchDaoAccount(
+    daoAddress: string
+  ): Promise<DaoAccount | undefined> {
+    return;
+  }
+
+  private async fetchDaoWithTokensFromState(
+    daoAccount: DaoAccount
+  ): Promise<DaoWithTokens | undefined> {
+    return;
+  }
+}
